CAMEL-10849: Salesforce: subscription channel ...
...created per component

Changes the `replayId` value from `Integer` to `Long`. The range of the
`replayId` is explicitly specified in the Salesforce documentation it's
safer to make it as big as can be.

(cherry picked from commit cdf97cb5223f555ea80fab5090afa1c167703e0a)

+++ components/camel-salesforce/camel-salesforce-component/src/main/java/org/apache/camel/component/salesforce/SalesforceEndpointConfig.java
@@ -154,9 +154,9 @@ public class SalesforceEndpointConfig implements Cloneable {
 
     // Streaming API properties
     @UriParam
-    private Integer defaultReplayId;
+    private Long defaultReplayId;
     @UriParam
-    private Map<String, Integer> initialReplayIdMap;
+    private Map<String, Long> initialReplayIdMap;
 
     // Salesforce Jetty9 HttpClient, set using reference
     @UriParam
@@ -587,7 +587,7 @@ public class SalesforceEndpointConfig implements Cloneable {
         return Collections.unmodifiableMap(valueMap);
     }
 
-    public Integer getDefaultReplayId() {
+    public Long getDefaultReplayId() {
         return defaultReplayId;
     }
 
@@ -595,18 +595,18 @@ public class SalesforceEndpointConfig implements Cloneable {
      * Default replayId setting if no value is found in {@link #initialReplayIdMap}
      * @param defaultReplayId
      */
-    public void setDefaultReplayId(Integer defaultReplayId) {
+    public void setDefaultReplayId(Long defaultReplayId) {
         this.defaultReplayId = defaultReplayId;
     }
 
-    public Map<String, Integer> getInitialReplayIdMap() {
+    public Map<String, Long> getInitialReplayIdMap() {
         return Optional.ofNullable(initialReplayIdMap).orElse(Collections.emptyMap());
     }
 
     /**
      * Replay IDs to start from per channel name.
      */
-    public void setInitialReplayIdMap(Map<String, Integer> initialReplayIdMap) {
+    public void setInitialReplayIdMap(Map<String, Long> initialReplayIdMap) {
         this.initialReplayIdMap = initialReplayIdMap;
     }
 }
